// src/app/types/Desafios.tsx
export type Desafio = {
   id: number;
   Titulo: string;
   descricao: string;
   pontos: number;
   concluido: boolean;
   multiplicador: number;
   Exigência_1: string;
   Exigência_2: string;
   Exigência_3: string;
   Exigência_4: string;
   tempo_da_atividade: number;
   links: string;
   tipo: string;
};